+++ src/main/java/domain/ladder/RandomScaffoldGenerator.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.ladder;
 
 public class RandomScaffoldGenerator implements ScaffoldGenerator {
