+++ src/main/java/domain/ladder/Scaffold.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.ladder;
 
 public enum Scaffold {
