+++ src/test/java/domain/ladder/ScaffoldTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.ladder;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DisplayNameGeneration;
